pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "discord-bot"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "Discord bot with various utility commands"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.0.0",
    "aiohttp>=3.8.0",
    "pytz>=2023.3"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "docstr-coverage>=2.3.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot with various utility commands"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.0"
aiohttp = "^3.8.0"
pytz = "^2023.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.0.0"
ruff = "^0.1.0"
black = "^23.0.0"
isort = "^5.12.0"
docstr-coverage = "^2.3.0"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyx?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
] 